requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
Flask-WTF==1.1.1
gunicorn==21.2.0
whitenoise==6.5.0 